config/requirements.txt
# Core dependencies
pandas>=1.5.0,<3.0.0
numpy>=1.21.0,<2.0.0
sqlalchemy>=2.0.0,<3.0.0
sqlglot>=10.0.0,<15.0.0

# Web framework
streamlit>=1.28.0,<2.0.0
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.20.0,<1.0.0
pydantic>=2.0.0,<3.0.0

# Language processing dependencies
transformers>=4.30.0,<5.0.0
torch>=2.0.0,<3.0.0
accelerate>=0.20.0,<1.0.0

# Database
duckdb>=0.8.0,<1.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
datasets>=2.14.0,<3.0.0
peft>=0.4.0,<1.0.0

# Evaluation dependencies
sqlparse>=0.4.0,<0.5.0
plotly>=5.0.0,<6.0.0
jinja2>=3.0.0,<4.0.0

# Production dependencies
gunicorn>=20.1.0,<22.0.0
psutil>=5.9.0,<6.0.0

